+++ vscode/CHANGELOG.md
@@ -4,6 +4,11 @@ Format:
 
 - VERSION [YYYY-MM-DD]: Summary.
 
+- 0.1.9 [2026-05-09]:	Testing...
+- 0.1.8 [2026-05-09]:	Testing...
+- 0.1.7 [2026-05-09]:	Testing...
+- 0.1.6 [2026-05-09]:	Badges in README.md
+- 0.1.5 [2026-05-09]:	Testing interaction with marketplace.
 - 0.1.4 [2026-05-05]:	More verbose output on activate.
 - 0.1.3 [2026-05-04]:	Prepare package.json for Azure/Marketplace
 - 0.1.2 [2026-05-02]:	Minor changes in tools